+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [v0.2.5] - 2026-06-24
+### Added
+- Refactor
+- Clash API secret
+
 ## [v0.2.4] - 2026-06-23
 ### Added
 - Remove duplicate outbounds from sing-box config